Add --reply-to feature.
This allows setting the Reply-To header in the sent emails.

That's useful if the email address from which the emails are sent
cannot be replied to. For example, if
  bot@example.com
sends the email to a mailing list
  list@example.com
and you want that people reply to the list and not to the bot,
use `--reply-to list@example.com`.